# torch.dropout: avoid unnecessary mul for p=1