@vadimkantorov 2020-11-07
One problem is that it changes the behavior from copying to no-copying. This can make some problems if there's an inplace op just after.