@kshitij12345 2020-11-08
@vadimkantorov 
> One problem is that it changes the behavior from copying to no-copying. This can make some problems if there's an inplace op just after only in train time, but maybe it's a minor consideration

Could you elaborate, cause I think the behaviour of inplace and copy would be preserved.