@heitorschueroff 2020-11-12
The only thing I'm concerned is that replacing `multiply` with `zero_tensor` will lose the sign on the zeros. The sign of zeros could change the result of atan2 for instance. @mruberry Can you advise if we want to preserve the sign of zeros here?