@heitorschueroff 2020-11-12
@kshitij12345 You are not synchronizing CUDA in your benchmark which is why you see the same results. You can refer to this [WIP benchmark recipe](https://deploy-preview-1176--pytorch-tutorials-preview.netlify.app/recipes/recipes/benchmark.html?highlight=benchmark) for a guide on benchmarking PyTorch code.