@ngimel 2020-11-13
`Timer` synchronizes cuda automatically, that said, 2us is extremely suspect.