@kshitij12345 2020-11-13
> The only thing I'm concerned is that replacing multiply with zero_tensor will lose the sign on the zeros. The sign of zeros could change the result of atan2 for instance. @mruberry Can you advise if we want to preserve the sign of zeros here?

Sample Snippet:
```python
>>> x = torch.randn(3,3)
>>> x
tensor([[ 2.2820, -0.9474,  0.1627],
        [ 1.0592,  0.4148, -0.5017],
        [-1.0321,  1.0442, -0.3100]])
>>> x * torch.zeros_like(x)
tensor([[0., -0., 0.],
        [0., 0., -0.],
        [-0., 0., -0.]])
>>> torch.zeros_like(x)
tensor([[0., 0., 0.],
        [0., 0., 0.],
        [0., 0., 0.]])

```