@ngimel 2020-11-13
tbh, I don't think this corner case deserves special treatment, so I'd be against merging this PR. However, there are a couple of real dropout issues that need fixing
1) with p=0 we still go to special cuda kernel instead of taking a shortcut
2) in the cuda kernel, both vectorized and regular, the comparisons are `<`, whereas given that curand produces 0-exclusive, 1-inclusive range they should be `<=`. 
Can you refocus this PR to fix these issues please?