@kshitij12345 2020-11-14
> with p=0 we still go to special cuda kernel instead of taking a shortcut

This I believe is handled properly, fused kernel is taken only when `p > 0 and p < 1`,
https://github.com/pytorch/pytorch/blob/fe7d1d7d0e0762cca6b323f6e2ad7cc252f8ccf3/aten/src/ATen/native/Dropout.cpp#L24-L26

Otherwise input is returned directly if `p==0`,
https://github.com/pytorch/pytorch/blob/fe7d1d7d0e0762cca6b323f6e2ad7cc252f8ccf3/aten/src/ATen/native/Dropout.cpp#L44-L47

For point 2,
> in the cuda kernel, both vectorized and regular, the comparisons are <, whereas given that curand produces 0-exclusive, 1-inclusive range they should be <=

Logically it makes sense, but with above condition `p > 0 and p < 1`, `p==1` never reaches those kernel.