@ngimel 2020-11-14
Yeah, my mistake, turns out that for p=0 we do take the shortcut and thus no values should be dropped. Changing < to <= still makes sense even for probabilities in (0,1), but that's fixing a very small skew, so maybe not worth the effort.